Refactor testing utilities into package
