+++ falcon_pachinko/testing/client.py
@@ -1,18 +1,28 @@
-"""Testing utilities for exercising websocket integrations."""
+"""High-level WebSocket test client for integration testing."""
 
 from __future__ import annotations
 
-import asyncio
 import dataclasses as dc
 import typing as typ
 from contextlib import asynccontextmanager
-from types import SimpleNamespace
 from urllib.parse import urlsplit
 
-import falcon.asgi
 import msgspec.json as msjson
 
-from .router import WebSocketRouter
+from ._common import (
+    Direction,
+    FrameKind,
+    PayloadKind,
+    MissingDependencyError,
+    _BINARY_PAYLOAD_REQUIRED_MSG,
+    _EXPECTED_BYTES_MSG,
+    _EXPECTED_TEXT_MSG,
+    _FAILED_JSON_DECODE_MSG,
+    _INSECURE_WEBSOCKET_MSG,
+    _MISSING_WEBSOCKETS_MSG,
+    _TEXT_PAYLOAD_REQUIRED_MSG,
+    _UNSUPPORTED_FRAME_KIND_MSG,
+)
 
 try:  # pragma: no cover - optional dependency exercised in tests
     from websockets.client import connect as _ws_connect
@@ -24,31 +34,6 @@ if typ.TYPE_CHECKING:  # pragma: no cover - typing only
 else:  # pragma: no cover - runtime fallback when dependency missing
     WebSocketClientProtocol = typ.Any  # type: ignore[misc,assignment]
 
-Direction = typ.Literal["send", "receive", "close", "error"]
-FrameKind = typ.Literal["text", "bytes", "json"]
-PayloadKind = typ.Literal["text", "bytes", "json", "close"]
-
-_MISSING_WEBSOCKETS_MSG = (
-    "WebSocketTestClient requires the 'websockets' package. Install "
-    "falcon-pachinko[testing] to enable these helpers."
-)
-_EXPECTED_TEXT_MSG = "Expected text frame but received bytes"
-_EXPECTED_BYTES_MSG = "Expected binary frame but received text"
-_TEXT_PAYLOAD_REQUIRED_MSG = "Text frames require str payloads"
-_BINARY_PAYLOAD_REQUIRED_MSG = "Binary frames require bytes payloads"
-_UNSUPPORTED_FRAME_KIND_MSG = "Unsupported frame kind: {frame_kind}"
-_FAILED_JSON_DECODE_MSG = "Failed to decode JSON payload: {message!r}"
-_JSON_FRAME_REQUIRED_MSG = "JSON frames must be text or binary payloads"
-_ORIGINAL_WS_RECEIVE_MSG = "Original websocket stub does not support receiving frames"
-_INSECURE_WEBSOCKET_MSG = (
-    "Insecure websocket URLs require allow_insecure=True. "
-    "Use a wss:// URL for secure connections."
-)
-
-
-class MissingDependencyError(RuntimeError):
-    """Raised when optional testing dependencies are unavailable."""
-
 
 @dc.dataclass(slots=True)
 class TraceEvent:
@@ -58,418 +43,6 @@ class TraceEvent:
     direction: Direction
     kind: PayloadKind
     payload: object
-
-
-class WebSocketSimulator:
-    """In-memory :class:`WebSocketLike` implementation for hermetic tests."""
-
-    def __init__(
-        self,
-        *,
-        inbound: asyncio.Queue[object] | None = None,
-        outbound: asyncio.Queue[object] | None = None,
-    ) -> None:
-        self._inbound = inbound or asyncio.Queue()
-        self._outbound = outbound or asyncio.Queue()
-        self._accepted = False
-        self._closed = False
-        self._close_code: int | None = None
-        self._subprotocol: str | None = None
-        self._json_encoder = msjson.Encoder()
-        self._default_decoder = msjson.Decoder()
-        self._decoders: dict[type[object], msjson.Decoder] = {}
-        self.sent_messages: list[object] = []
-        self.received_messages: list[object] = []
-
-    @property
-    def accepted(self) -> bool:
-        """Return ``True`` once :meth:`accept` has been called."""
-        return self._accepted
-
-    @property
-    def closed(self) -> bool:
-        """Return ``True`` once :meth:`close` has been called."""
-        return self._closed
-
-    @property
-    def close_code(self) -> int | None:
-        """Return the code provided to :meth:`close`, if any."""
-        return self._close_code
-
-    @property
-    def subprotocol(self) -> str | None:
-        """Return the negotiated subprotocol, if any."""
-        return self._subprotocol
-
-    def pending_inbound(self) -> int:
-        """Return the number of queued inbound frames."""
-        return self._inbound.qsize()
-
-    def pending_outbound(self) -> int:
-        """Return the number of queued outbound frames."""
-        return self._outbound.qsize()
-
-    def _decoder_for(self, payload_type: type[object] | None) -> msjson.Decoder:
-        """Return a cached decoder for ``payload_type``."""
-        if payload_type is None:
-            return self._default_decoder
-        decoder = self._decoders.get(payload_type)
-        if decoder is None:
-            decoder = msjson.Decoder(payload_type)
-            self._decoders[payload_type] = decoder
-        return decoder
-
-    async def accept(self, subprotocol: str | None = None) -> None:
-        """Record that the handshake was accepted."""
-        self._accepted = True
-        self._subprotocol = subprotocol
-
-    async def close(self, code: int = 1000) -> None:
-        """Record that the connection was closed."""
-        self._closed = True
-        self._close_code = code
-
-    async def send_media(self, data: object) -> None:
-        """Record ``data`` as an outbound frame."""
-        await self._outbound.put(data)
-        self.sent_messages.append(data)
-
-    async def receive_media(self) -> object:
-        """Return the next inbound frame queued via :meth:`push_message`."""
-        message = await self._inbound.get()
-        self.received_messages.append(message)
-        return message
-
-    async def next_sent(self) -> object:
-        """Await the next outbound frame emitted by the simulator."""
-        return await self._outbound.get()
-
-    def pop_sent(self) -> object:
-        """Pop the next outbound frame synchronously."""
-        try:
-            return self._outbound.get_nowait()
-        except asyncio.QueueEmpty as exc:
-            msg = "No sent messages available"
-            raise LookupError(msg) from exc
-
-    async def send_text(self, message: str) -> None:
-        """Send a UTF-8 text frame."""
-        if not isinstance(message, str):
-            raise TypeError(_TEXT_PAYLOAD_REQUIRED_MSG)
-        await self.send_media(message)
-
-    async def send_bytes(self, payload: bytes | bytearray | memoryview) -> None:
-        """Send a binary frame."""
-        if not isinstance(payload, bytes | bytearray | memoryview):
-            raise TypeError(_BINARY_PAYLOAD_REQUIRED_MSG)
-        await self.send_media(bytes(payload))
-
-    async def send_json(self, payload: object) -> None:
-        """Encode ``payload`` as JSON and send it as bytes."""
-        await self.send_media(self._json_encoder.encode(payload))
-
-    async def receive_text(self) -> str:
-        """Receive the next frame ensuring it is textual."""
-        message = await self.receive_media()
-        if not isinstance(message, str):
-            raise TypeError(_EXPECTED_TEXT_MSG)
-        return message
-
-    async def receive_bytes(self) -> bytes:
-        """Receive the next frame ensuring it is binary."""
-        message = await self.receive_media()
-        if isinstance(message, bytes):
-            return message
-        raise TypeError(_EXPECTED_BYTES_MSG)
-
-    async def receive_json(self, payload_type: type[object] | None = None) -> object:
-        """Receive and decode a JSON payload."""
-        message = await self.receive_media()
-        if isinstance(message, str):
-            data = message.encode("utf-8")
-        elif isinstance(message, bytes | bytearray | memoryview):
-            data = bytes(message)
-        else:
-            raise TypeError(_FAILED_JSON_DECODE_MSG.format(message=message))
-        decoder = self._decoder_for(payload_type)
-        return decoder.decode(data)
-
-    async def push_message(self, payload: object, *, kind: FrameKind = "json") -> None:
-        """Queue ``payload`` as if it were received from the peer."""
-        data = self._prepare_inbound_payload(payload, kind)
-        await self._inbound.put(data)
-
-    def _prepare_inbound_payload(self, payload: object, kind: FrameKind) -> object:
-        if kind == "text":
-            return self._prepare_text_payload(payload)
-        if kind == "bytes":
-            return self._prepare_bytes_payload(payload)
-        if kind == "json":
-            return self._json_encoder.encode(payload)
-        raise ValueError(
-            _UNSUPPORTED_FRAME_KIND_MSG.format(frame_kind=kind)
-        )  # pragma: no cover - safeguarded by FrameKind literal
-
-    def _prepare_text_payload(self, payload: object) -> str:
-        if not isinstance(payload, str):
-            raise TypeError(_TEXT_PAYLOAD_REQUIRED_MSG)
-        return payload
-
-    def _prepare_bytes_payload(self, payload: object) -> bytes:
-        if not isinstance(payload, bytes | bytearray | memoryview):
-            raise TypeError(_BINARY_PAYLOAD_REQUIRED_MSG)
-        return bytes(payload)
-
-    async def push_text(self, message: str) -> None:
-        """Queue a UTF-8 text frame."""
-        await self.push_message(message, kind="text")
-
-    async def push_bytes(self, payload: bytes | bytearray | memoryview) -> None:
-        """Queue a binary frame."""
-        await self.push_message(payload, kind="bytes")
-
-    async def push_json(self, payload: object) -> None:
-        """Queue a JSON payload."""
-        await self.push_message(payload, kind="json")
-
-    @asynccontextmanager
-    async def connected(
-        self, *, subprotocol: str | None = None
-    ) -> typ.AsyncIterator[WebSocketSimulator]:
-        """Accept the connection on entry and close it on exit."""
-        await self.accept(subprotocol=subprotocol)
-        try:
-            yield self
-        finally:
-            if not self._closed:
-                await self.close()
-
-
-@dc.dataclass(slots=True)
-class SimulatorConnection:
-    """Describe a simulated connection managed by the pytest harness."""
-
-    path: str
-    router: WebSocketRouter
-    simulator: WebSocketSimulator
-    request: object
-    websocket: _OriginalWebSocket
-    _json_decoder: msjson.Decoder = dc.field(init=False, repr=False)
-
-    def __post_init__(self) -> None:
-        """Initialize a decoder used to inspect outbound JSON frames."""
-        self._json_decoder = msjson.Decoder()
-
-    @property
-    def accepted(self) -> bool:
-        """Return ``True`` when the simulator accepted the handshake."""
-        return self.simulator.accepted
-
-    @property
-    def closed(self) -> bool:
-        """Return ``True`` once the simulator recorded connection closure."""
-        return self.simulator.closed
-
-    @property
-    def close_code(self) -> int | None:
-        """Expose the close code recorded by the simulator."""
-        return self.simulator.close_code
-
-    @property
-    def sent_messages(self) -> list[object]:
-        """Return a snapshot of frames emitted by the resource."""
-        return list(self.simulator.sent_messages)
-
-    def pop_sent(self) -> object:
-        """Pop the next outbound frame without decoding."""
-        return self.simulator.pop_sent()
-
-    def pop_sent_json(self, payload_type: type[object] | None = None) -> object:
-        """Pop the next outbound frame and decode it as JSON."""
-        raw = self.pop_sent()
-        if isinstance(raw, str):
-            data = raw.encode("utf-8")
-        elif isinstance(raw, bytes | bytearray | memoryview):
-            data = bytes(raw)
-        else:  # pragma: no cover - safeguarded by simulator helpers
-            raise TypeError(_JSON_FRAME_REQUIRED_MSG)
-        decoder = (
-            self._json_decoder if payload_type is None else msjson.Decoder(payload_type)
-        )
-        return decoder.decode(data)
-
-    async def push_json(self, payload: object) -> None:
-        """Queue a JSON payload for the resource to consume."""
-        await self.simulator.push_json(payload)
-
-    async def push_text(self, message: str) -> None:
-        """Queue a UTF-8 text frame for the resource."""
-        await self.simulator.push_text(message)
-
-    async def push_bytes(self, payload: bytes | bytearray | memoryview) -> None:
-        """Queue a binary frame for the resource."""
-        await self.simulator.push_bytes(payload)
-
-
-class _OriginalWebSocket:
-    """Minimal stub representing the ASGI-provided websocket."""
-
-    def __init__(self) -> None:
-        self.accepted = False
-        self.closed = False
-        self.close_code: int | None = None
-        self.subprotocol: str | None = None
-        self.sent: list[object] = []
-
-    async def accept(self, subprotocol: str | None = None) -> None:
-        self.accepted = True
-        self.subprotocol = subprotocol
-
-    async def close(self, code: int = 1000) -> None:
-        self.closed = True
-        self.close_code = code
-
-    async def send_media(self, data: object) -> None:  # pragma: no cover - unused
-        self.sent.append(data)
-
-    async def receive_media(self) -> object:  # pragma: no cover - unused
-        raise RuntimeError(_ORIGINAL_WS_RECEIVE_MSG)
-
-
-class _HarnessSimulator(WebSocketSimulator):
-    """Simulator variant that mirrors lifecycle events to the original stub."""
-
-    def __init__(self) -> None:
-        super().__init__()
-        self._original: _OriginalWebSocket | None = None
-
-    def bind_original(self, original: _OriginalWebSocket) -> None:
-        """Associate ``original`` so lifecycle events stay in sync."""
-        self._original = original
-
-    async def accept(self, subprotocol: str | None = None) -> None:
-        await super().accept(subprotocol=subprotocol)
-        if self._original is not None and not self._original.accepted:
-            await self._original.accept(subprotocol=subprotocol)
-
-    async def close(self, code: int = 1000) -> None:
-        await super().close(code)
-        if self._original is not None and not self._original.closed:
-            await self._original.close(code)
-
-
-@dc.dataclass(slots=True)
-class _TestRequest:
-    """Lightweight stand-in for :class:`falcon.Request`."""
-
-    path: str
-    path_template: str
-    context: SimpleNamespace = dc.field(default_factory=SimpleNamespace)
-
-
-class SimulatorRouterHarness:
-    """Manage a simulator-backed router mounted on a Falcon ASGI app."""
-
-    def __init__(self, *, mount: str = "/") -> None:
-        self.app = falcon.asgi.App()
-        self._mount_prefix = self._normalize_mount(mount)
-        self._pending_simulator: WebSocketSimulator | None = None
-        self.router = WebSocketRouter(simulator_factory=self._provide_simulator)
-        self._mounted = False
-        self.mount(self._mount_prefix)
-
-    def mount(self, prefix: str) -> None:
-        """Mount the router at ``prefix`` and register it with the app."""
-        normalized = self._normalize_mount(prefix)
-        if self._mounted:
-            if normalized != self._mount_prefix:
-                msg = f"router already mounted at {self._mount_prefix!r}"
-                raise RuntimeError(msg)
-            return
-        self.router.mount(normalized)
-        self.app.add_route(normalized, self.router)
-        self._mount_prefix = normalized
-        self._mounted = True
-
-    def _normalize_mount(self, prefix: str) -> str:
-        if not prefix:
-            return "/"
-        if not prefix.startswith("/"):
-            prefix = f"/{prefix}"
-        return prefix.rstrip("/") or "/"
-
-    def _provide_simulator(self, req: object, ws: object) -> WebSocketSimulator:
-        """Return the simulator associated with the next connection."""
-        simulator = self._pending_simulator or _HarnessSimulator()
-        self._pending_simulator = None
-        if isinstance(simulator, _HarnessSimulator) and isinstance(
-            ws, _OriginalWebSocket
-        ):
-            simulator.bind_original(ws)
-        return simulator
-
-    def _compose_path(self, path: str) -> str:
-        if not path:
-            return self._mount_prefix
-        if not path.startswith("/"):
-            path = f"/{path}"
-        if path.startswith(self._mount_prefix):
-            return path
-        if self._mount_prefix == "/":
-            return path
-        return f"{self._mount_prefix}{path}"
-
-    @asynccontextmanager
-    async def connect(
-        self,
-        path: str,
-        *,
-        initial_inbound: typ.Iterable[tuple[object, FrameKind]] | None = None,
-    ) -> typ.AsyncIterator[SimulatorConnection]:
-        """Dispatch ``path`` through the router yielding a connection helper."""
-        if not self._mounted:
-            msg = "router must be mounted before establishing connections"
-            raise RuntimeError(msg)
-
-        simulator = _HarnessSimulator()
-        self._pending_simulator = simulator
-        if initial_inbound is not None:
-            for payload, kind in initial_inbound:
-                await simulator.push_message(payload, kind=kind)
-        request_path = self._compose_path(path)
-        request = _TestRequest(path=request_path, path_template=self._mount_prefix)
-        original = _OriginalWebSocket()
-        try:
-            await self.router.on_websocket(request, original)
-            connection = SimulatorConnection(
-                path=request_path,
-                router=self.router,
-                simulator=simulator,
-                request=request,
-                websocket=original,
-            )
-            yield connection
-        finally:
-            if not simulator.closed:
-                await simulator.close()
-            if not original.closed:
-                await original.close()
-
-
-try:  # pragma: no cover - optional dependency for fixture registration
-    import pytest
-except ImportError:  # pragma: no cover - fixture only available under pytest
-    pytest = None  # type: ignore[assignment]
-else:
-
-    @pytest.fixture
-    def websocket_simulator() -> typ.Iterator[SimulatorRouterHarness]:
-        """Provide a router harness pre-wired with a simulator factory."""
-        harness = SimulatorRouterHarness()
-        try:
-            yield harness
-        finally:
-            harness._pending_simulator = None
 
 
 class WebSocketSession:
@@ -872,13 +445,3 @@ class WebSocketTestClient:
             finally:
                 if not session.closed:
                     await session.close()
-
-
-__all__ = [
-    "MissingDependencyError",
-    "SimulatorConnection",
-    "SimulatorRouterHarness",
-    "TraceEvent",
-    "WebSocketSimulator",
-    "WebSocketTestClient",
-]
